objective of this project was to utilize Power BI to analyze and visualize COVID-19 data
The COVID-19 pandemic has had a significant impact globally, necessitating the need for robust data analysis to understand and manage the spread of the virus. This project involves the comprehensive analysis of COVID-19 cases using Microsoft Power BI, providing insights into the spread, impact, and trends of the virus over time.